Rename reconciliation columns file
